Remove `lab` tag.
Remove `lab` tag. Now the new form component is in use.
