+++ .storybook/config.js
@@ -2,7 +2,6 @@ import { configure } from '@kadira/storybook';
 
 function loadStories() {
   require('../stories');
-  require('../stories/form.stories.js');
   require('../stories/modal.stories.js');
   require('../stories/table.stories.js');
   require('../stories/datepicker.stories.js')
@@ -12,14 +11,14 @@ function loadStories() {
   require('../stories/animator.stories.js');
   require('../stories/stepper.stories.js');
   require('../stories/pagination.stories.js');
-  require('../stories/input-in-lab.stories.js');
-  require('../stories/form-in-lab.stories.js');
+  require('../stories/input.stories.js');
+  require('../stories/form.stories.js');
   require('../stories/checkbox.stories.js');
   require('../stories/radio.stories.js');
   require('../stories/select.stories.js');
   require('../stories/switch.stories.js');
   require('../stories/numberInput.stories.js');
-  require('../stories/datepicker-in-lab.stories.js');
+  require('../stories/datepicker.stories.js');
 }
 
 configure(loadStories, module);
