--- components/NumberInput.jsx
@@ -1,118 +1,0 @@
-import React, { Component, PropTypes } from 'react';
-import classNames from 'classnames';
-import { register } from './highorder/FormItem';
-import '../style/numberinput.scss';
-
-const propTypes = {
-  value: PropTypes.number,
-  min: PropTypes.number,
-  max: PropTypes.number,
-  step: PropTypes.number.isRequired,
-  onChange: PropTypes.func,
-};
-
-const defaultProps = {
-  step: 1,
-};
-
-class NumberInput extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { focus: false, value: this.props.value };
-    this.handleDownButtonClick = this.handleDownButtonClick.bind(this);
-    this.handleUpButtonClick = this.handleUpButtonClick.bind(this);
-  }
-
-  componentWillReceiveProps(props) {
-    this.setState({ value: props.value });
-  }
-
-  parseValue(value) {
-    let val = value;
-    if (val === '') {
-      val = '';
-    } else if (!Number.isNaN(val)) {
-      val = Number.parseInt(val, 10);
-    }
-  }
-
-  handleDownButtonClick() {
-    const { step, min } = this.props;
-    const v = this.state.value ? Number.parseInt(this.state.value, 10) : 0;
-    const resultValue = (v - step) < min ? min : v - step;
-    if (this.props.onChange) {
-      this.props.onChange(resultValue);
-    }
-    this.setState({ value: resultValue, focus: true });
-    this.numberInput.focus();
-  }
-
-  handleUpButtonClick() {
-    const { step, max } = this.props;
-    const v = this.state.value ? Number.parseInt(this.state.value, 10) : 0;
-    const resultValue = (v + step) > max ? max : v + step;
-    if (this.props.onChange) {
-      this.props.onChange(resultValue);
-    }
-    this.setState({ value: resultValue, focus: true });
-    this.numberInput.focus();
-  }
-
-  render() {
-    const wrapperClassName = this.state.focus ?
-      'zc-number-input-wrapper-active' : 'zc-number-input-wrapper';
-    const { value } = this.state;
-    const { min, max } = this.props;
-    const inputDownClassName = classNames({
-      'zc-number-input-down': true,
-      'zc-number-input-button-disabled': min === value,
-    });
-    const inputUpClassName = classNames({
-      'zc-number-input-up': true,
-      'zc-number-input-button-disabled': max === value,
-    });
-    return (
-      <div className={wrapperClassName}>
-        <div
-          className={inputDownClassName}
-          onClick={this.handleDownButtonClick}
-        >
-          -
-        </div>
-        <input
-          className="zc-number-input"
-          type="text"
-          value={this.state.value}
-          onChange={(e) => {
-            const v = e.target.value;
-            if (!/^[-+]?\d*$/.test(v)) {
-              return;
-            }
-            if (this.props.onChange) {
-              this.props.onChange(v);
-            }
-            this.setState({ value: v });
-          }}
-          onFocus={() => {
-            this.setState({ focus: true });
-          }}
-          onBlur={() => {
-            this.setState({ focus: false });
-          }}
-          ref={(ref) => { this.numberInput = ref; }}
-        />
-        <div
-          className={inputUpClassName}
-          onClick={this.handleUpButtonClick}
-        >
-          +
-        </div>
-      </div>
-    );
-  }
-}
-
-NumberInput.propTypes = propTypes;
-NumberInput.defaultProps = defaultProps;
-
-export default register(NumberInput);
