+++ utils/validate/validator/number.js
@@ -1,22 +1,24 @@
 import isNumber from '../../number/isNumber';
 import rules from '../rule';
+import format from '../../format';
 
 // rule is a single rule
 export default function number(rule, value, callback, fields, options) {
   const errors = [];
-  rules.required(rule, value, callback, fields, errors, options);
   const v = Number.parseFloat(value);
-  if (!isNumber(v)) {
-    errors.push('你需要提供一个数字');
+  if (value === undefined || value === null || value === '' || Number.isNaN(value)) {
+    rules.required(rule, value, callback, errors, fields, options);
+  } else if (!isNumber(v)) {
+    errors.push(options.message.number.default);
   } else {
     if ('min' in rule) {
       if (v < rule.min) {
-        errors.push('太小');
+        errors.push(format(options.message.number.min, rule.fieldName, rule.min));
       }
     }
     if ('max' in rule) {
       if (v > rule.max) {
-        errors.push('太大');
+        errors.push(format(options.message.number.max, rule.fieldName, rule.max));
       }
     }
   }
