+++ utils/validate/validator/string.js
@@ -1,18 +1,27 @@
 import isString from '../../string/isString';
+import rules from '../rule';
+import format from '../../format';
 
-export default function array(rule, value, callback, fields) {
+/**
+ * The logic is when the value is '', null, undefined, then show required
+ * If not a string provided, show need a string.
+ * If the string is provided, then you can just test it.
+ */
+export default function string(rule, value, callback, fields, options) {
   const errors = [];
-  if (!isString(value)) {
-    errors.push('必须提供一个字符串');
+  if (value === undefined || value === null || value === '') {
+    rules.required(rule, value, callback, errors, fields, options);
+  } else if (!isString(value)) {
+    errors.push(options.message.default);
   } else {
     if ('min' in rule) {
       if (value.length < rule.min) {
-        errors.push('太小了');
+        errors.push(format(options.message.string.min, rule.fieldName, rule.min));
       }
     }
     if ('max' in rule) {
       if (value.length > rule.max) {
-        errors.push('太大了');
+        errors.push(format(options.message.string.max, rule.fieldName, rule.max));
       }
     }
   }
